Patch 0.06
= Changes to inner code to exclude spheres.
= Changed project settings to fit the testing scenario

Bit of context here, the testing results, i did not trust them, and i was right in doing so. This is due to the fact that after deepening my research, the testing enviroments were not accurate with eachother. However, due to this, the sample size will be decreased due to the fact that i have already done 18 simulations. This will be reduced to 12 instead.

Spheres were also removed, due to the fact that it seems like Godot's physics systems are, wrong??? For some reason, the spheres just go through walls without rhyme or reason. After a full day of testing, messing with editor values, googling and asking the Godot community (They ignored me) no sollution was found, and i had to exclude it from the testing scenario's (Squares did work properly)